Add GCP functionality to _scheme_rmvm_keepdisk_rollback. + Add disk labelling for GCP. + Fix for _scheme_addnewvm_rmdisk_rollback when using myhosttypes
