+++ redeploy/_scheme_rmvm_keepdisk_rollback/tasks/_add_src_diskinfo_to_cluster_hosts_target__aws.yml
@@ -18,10 +18,10 @@
     cluster_hosts_target: |
       {%- for cht_host in cluster_hosts_target -%}
         {%- for cht_autovol in cht_host.auto_volumes -%}
-          {%- for chs_host_info_result in r__ec2_instance_info.instances | selectattr('tags.lifecycle_state', '!=', 'current') -%}
-            {%- if cht_host.hostname | regex_replace('-(?!.*-).*') == chs_host_info_result.tags.Name | regex_replace('-(?!.*-).*') -%}
-              {%- for chs_host_diskinfo in chs_host_info_result.block_device_mappings | selectattr('device_name', '==', cht_autovol.device_name) -%}
-                {%- set _ = cht_autovol.update({'src': {'instance_id': chs_host_info_result.instance_id, 'device_name': chs_host_diskinfo.device_name, 'volume_id': chs_host_diskinfo.ebs.volume_id }}) -%}
+          {%- for ec2_instance_info_result in r__ec2_instance_info.instances | selectattr('tags.lifecycle_state', '!=', 'current') -%}
+            {%- if cht_host.hostname | regex_replace('-(?!.*-).*') == ec2_instance_info_result.tags.Name | regex_replace('-(?!.*-).*') -%}
+              {%- for chs_host_diskinfo in ec2_instance_info_result.block_device_mappings | selectattr('device_name', '==', cht_autovol.device_name) -%}
+                {%- set _ = cht_autovol.update({'src': {'instance_id': ec2_instance_info_result.instance_id, 'device_name': chs_host_diskinfo.device_name, 'volume_id': chs_host_diskinfo.ebs.volume_id }}) -%}
               {%- endfor -%}
             {%- endif -%}
           {%- endfor -%}
