+++ redeploy/tasks/main.yml
@@ -3,7 +3,7 @@
 - name: Preflight check - Redeploy
   block:
     - assert: { that: "clean is not defined", msg: "Must not set the 'clean' variable for a redeploy" }
-    - assert: { that: "canary is defined and (canary is defined and canary in ['start', 'finish', 'none', 'tidy', 'revert'])", msg: "Canary must be 'start', 'finish', 'none', 'tidy' or 'revert'" }
+    - assert: { that: "canary is defined and (canary is defined and canary in ['start', 'finish', 'none', 'tidy'])", msg: "Canary must be 'start', 'finish', 'none' or 'tidy'" }
     - assert: { that: "redeploy_scheme is defined and redeploy_scheme in redeploy_schemes_supported" }
     - assert: { that: "cluster_hosts_state | length", msg: "Redeploy only possible with an existing cluster." }
